#15: package renaming to match new group id
