--- src/main/kotlin/com/github/helpermethod/ArtifactoryExtensions.kt
@@ -1,9 +1,0 @@
-package com.github.helpermethod
-
-import org.apache.maven.artifact.Artifact
-import org.apache.maven.artifact.factory.ArtifactFactory
-import org.apache.maven.artifact.versioning.VersionRange.createFromVersionSpec
-import org.apache.maven.model.Dependency
-
-fun ArtifactFactory.createDependencyArtifact(dependency: Dependency): Artifact =
-    createDependencyArtifact(dependency.groupId, dependency.artifactId, createFromVersionSpec(dependency.version), dependency.type, dependency.classifier, dependency.scope, dependency.optional)
