--- src/main/kotlin/com/github/helpermethod/DependencyExtensions.kt
@@ -1,6 +1,0 @@
-package com.github.helpermethod
-
-import org.apache.maven.model.Dependency
-
-fun Dependency.isConcrete() =
-    version != null && !version.contains("\\$\\{[^}]+}".toRegex())
