--- src/main/kotlin/com/github/helpermethod/GitProvider.kt
@@ -1,11 +1,0 @@
-package com.github.helpermethod
-
-interface GitProvider : AutoCloseable {
-
-    fun hasRemoteBranch(remoteBranchName: String) : Boolean
-
-    fun checkoutNewBranch(branchName: String)
-    fun add(filePattern: String)
-    fun commit(author: String, message: String)
-    fun push(localBranchName: String)
-}
