+++ src/main/kotlin/io/github/georgberky/maven/plugins/depsupdate/DependencyExtensions.kt
@@ -1,0 +1,6 @@
+package io.github.georgberky.maven.plugins.depsupdate
+
+import org.apache.maven.model.Dependency
+
+fun Dependency.isConcrete() =
+    version != null && !version.contains("\\$\\{[^}]+}".toRegex())
