+++ src/main/kotlin/io/github/georgberky/maven/plugins/depsupdate/GitProvider.kt
@@ -1,0 +1,11 @@
+package io.github.georgberky.maven.plugins.depsupdate
+
+interface GitProvider : AutoCloseable {
+
+    fun hasRemoteBranch(remoteBranchName: String) : Boolean
+
+    fun checkoutNewBranch(branchName: String)
+    fun add(filePattern: String)
+    fun commit(author: String, message: String)
+    fun push(localBranchName: String)
+}
