+++ src/main/kotlin/io/github/georgberky/maven/plugins/depsupdate/Update.kt
@@ -1,0 +1,5 @@
+package io.github.georgberky.maven.plugins.depsupdate
+
+import org.jsoup.nodes.Document
+
+data class Update(val groupId: String, val artifactId: String, val version: String, val latestVersion: String, val modification: (Document) -> Unit)
