--- src/test/kotlin/com/github/helpermethod/ParentVersionUpdateTest.kt
@@ -1,18 +1,0 @@
-package com.github.helpermethod
-
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-
-internal class ParentVersionUpdateTest {
-
-    @Test
-    fun updatedPom() {
-        val pom = readPom("src/test/resources/poms/parentPom.xml")
-
-        val updateVersionUnderTest = ParentVersionUpdate("dummy", "dummy", "dummy", "0.7.0", pom)
-
-        val updatedPom = updateVersionUnderTest.updatedPom()
-
-        assertThat(extractFromPom(updatedPom, "project.parent.version")).isEqualTo("0.7.0")
-    }
-}
