--- src/test/resources-its/com/github/helpermethod/UpdateMojoNativeGitIT/nativeProviderIsSet/pom.xml
@@ -1,37 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>com.github.helpermethod</groupId>
-    <artifactId>native-git-provider-test</artifactId>
-    <version>0.6.0</version>
-
-    <scm>
-        <connection>someConnection</connection>
-        <developerConnection>someDeveloperConnection</developerConnection>
-    </scm>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter</artifactId>
-            <version>5.5.2</version>
-        </dependency>
-        <dependency>
-            <groupId>org.assertj</groupId>
-            <artifactId>assertj-core</artifactId>
-            <version>3.15.0</version>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>${project.groupId}</groupId>
-                <artifactId>${project.artifactId}</artifactId>
-                <version>${project.version}</version>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
