+++ src/test/resources/poms/parentPom.xml
@@ -3,12 +3,12 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.github.helpermethod</groupId>
+        <groupId>io.github.georgberky.maven.plugins.depsupdate</groupId>
         <artifactId>dependency-update-parent</artifactId>
         <version>0.6.0</version>
     </parent>
 
-    <groupId>com.github.helpermethod</groupId>
+    <groupId>io.github.georgberky.maven.plugins.depsupdate</groupId>
     <artifactId>dependency-update-maven-plugin</artifactId>
     <version>0.6.0</version>
